Some file moving and improvments
