+++ src/server/api-handler.js
@@ -1,22 +1,19 @@
 import { Router } from 'express'
 import { json, urlencoded } from 'body-parser'
-import logger from './logger'
-import authRequired from './authRequired'
+import logger from './helpers/logger'
+import errorHandler from './helpers/error-handler'
 
 import UsersController from './controllers/users-controller'
 import AuthController from './controllers/auth-controller'
 
 const APIHandler = new Router()
+
+APIHandler.use(logger)
 APIHandler.use(json())
 APIHandler.use(urlencoded({extended: true}))
-APIHandler.use(logger)
+APIHandler.use(errorHandler)
 
 APIHandler.use('/users', UsersController)
 APIHandler.use('/auth', AuthController)
-
-APIHandler.use('/test', authRequired)
-APIHandler.get('/test', (req, res) => {
-  res.json(req.currentUser)
-})
 
 export default APIHandler
